feat: numerous i915 additions

+++ lact-schema/src/lib.rs
@@ -169,13 +169,13 @@ pub struct NvidiaClocksTable {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq, Eq)]
 pub struct IntelClocksTable {
-    pub gt_freq: Option<(u32, u32)>,
+    pub gt_freq: Option<(u64, u64)>,
     /// - rpn_freq: The Render Performance (RP) N level, which is the minimal one.
-    pub rpn_freq: Option<u32>,
+    pub rpn_freq: Option<u64>,
     /// - rpe_freq: The Render Performance (RP) E level, which is the efficient one.
-    pub rpe_freq: Option<u32>,
+    pub rpe_freq: Option<u64>,
     /// - rp0_freq: The Render Performance (RP) 0 level, which is the maximum one.
-    pub rp0_freq: Option<u32>,
+    pub rp0_freq: Option<u64>,
 }
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone)]
@@ -288,7 +288,7 @@ pub struct PmfwInfo {
 pub struct ClockspeedStats {
     pub gpu_clockspeed: Option<u64>,
     /// Target clock
-    pub current_gfxclk: Option<u16>,
+    pub current_gfxclk: Option<u64>,
     pub vram_clockspeed: Option<u64>,
 }
